Testing Branch

+++ server/routes/campgrounds.js
@@ -34,8 +34,12 @@ router.put("/:id", async (req, res) => {
 });
 
 router.delete("/:id", async (req, res) => {
-  const deleteCamp = await Campground.findByIdAndDelete(req.params.id);
-  res.status(200).send("Deleted successful");
+  if (!req.isAuthenticated()) {
+    res.status(401).send("You must sign in first.");
+  } else {
+    const deleteCamp = await Campground.findByIdAndDelete(req.params.id);
+    res.status(200).send("Deleted successful");
+  }
 });
 
 router.delete("/:id/reviews/:reviewid", async (req, res) => {
@@ -46,7 +50,6 @@ router.delete("/:id/reviews/:reviewid", async (req, res) => {
 });
 
 router.post("/create", async (req, res) => {
-  console.log("New Campground Created");
   const camp = new Campground(req.body);
   await camp.save();
   res.status(200).send("Created successful");
